tools: rename gst-discoverer binary to gst-discoverer-0.10
We're not providing a wrapper like we do for the tools in core,
since wrappers are confusing (e.g. for debugging purposes),
mostly pointless (since the API is likely to change between
major versions), and cause packaging issues when packages for
two different major versions are to be installed in parallel.

https://bugzilla.gnome.org/show_bug.cgi?id=633023