xviddec: bodge to avoid crashes
It seems xvidcore overreads its input buffer, so a nasty workaround
is to allocate some more memory (16 bytes seem to be enough).
There is no apparent image corruption with these extra bytes set to 0,
valgrind is much happier, and the crashes go away.
It is ugly, and slower though. But then, xviddec is currently
not autoplugged for playback anyway.

https://bugzilla.gnome.org/show_bug.cgi?id=334107