kate: ensure the kate pad does not shoot ahead of the video pad
Sync both pads by waiting in the kate chain function.
Do not reset our internal segment from segment updates, in order
to be able to map video running time to kate running time, to
give libtiger the timestamp it expects. This allows us to use
running time to sync to video, which is The Right Way.

https://bugzilla.gnome.org/show_bug.cgi?id=600929